Move media session callback object to another file
